Split controllers/firmware.py into a subpackage
Per #123. ``controllers/firmware.py`` (~1700 lines) mixed the job
queue, the ``esphome`` subprocess wrapper, error-pattern detection,
and the WS handler glue. Split into a subpackage so each surface
is reviewable on its own:

- ``controllers/firmware/__init__.py`` — re-exports
  ``FirmwareController`` so existing
  ``from .controllers.firmware import FirmwareController`` keeps
  resolving.
- ``controllers/firmware/constants.py`` — error-pattern regexes,
  job-state strings, queue/cleanup tunables.
- ``controllers/firmware/helpers.py`` — pure free functions
  (``_find_esphome_cmd``, ``_trim_job_output``, ``_mark_job_terminal``,
  ``_names_touched_by_job``, ``_parse_progress``,
  ``_signal_process_group``, ``_terminate_subtree_windows``,
  ``_validate_port``, ``_verify_esphome_importable``,
  ``_is_no_module_named_esphome``).
- ``controllers/firmware/controller.py`` — ``FirmwareController``
  itself + the queue runner.

No behaviour change. Two production callers
(``controllers/devices.py``, ``controllers/editor.py``) updated to
import ``_find_esphome_cmd`` from the new ``firmware.helpers`` module
directly rather than going through ``__init__.py`` — keeps the
package surface honest about what's part of the public API
(only ``FirmwareController``) and what's an internal helper.

Tests mirror the production split: ``tests/test_firmware_*.py``
moved to ``tests/controllers/firmware/test_*.py`` (private-symbol
imports updated to point at ``firmware.helpers`` /
``firmware.constants``). All 701 tests pass; pre-commit clean.
