+++ tests/controllers/__init__.py
@@ -1,0 +1,1 @@
+"""Per-controller test packages mirror ``esphome_device_builder/controllers/``."""
