+++ tests/controllers/firmware/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for ``controllers/firmware/`` — split mirrors the production subpackage."""
