+++ tests/controllers/firmware/test_address_cache.py
