+++ tests/controllers/firmware/test_follow_jobs_race.py
