+++ tests/controllers/firmware/test_job_output_cap.py
@@ -15,13 +15,13 @@ process OOMs first.
 
 from __future__ import annotations
 
-from esphome_device_builder.controllers.firmware import (
+from esphome_device_builder.controllers.firmware.constants import (
     _INFLIGHT_TRIM_KEEP,
     _MAX_OUTPUT_LINES_INFLIGHT,
     _MAX_OUTPUT_LINES_RETAINED,
     _OUTPUT_TRIM_NOTICE_PREFIX,
-    _trim_job_output,
 )
+from esphome_device_builder.controllers.firmware.helpers import _trim_job_output
 from esphome_device_builder.models import FirmwareJob, JobStatus, JobType
 
 
@@ -119,7 +119,7 @@ def test_is_no_module_named_esphome_matches_exact_quoted_form() -> None:
     different reasons) would tell users to reinstall ESPHome
     itself when the real fix is to install a different dependency.
     """
-    from esphome_device_builder.controllers.firmware import _is_no_module_named_esphome
+    from esphome_device_builder.controllers.firmware.helpers import _is_no_module_named_esphome
 
     # CPython's exact ModuleNotFoundError emission — the format we
     # actually need to detect.
@@ -158,8 +158,8 @@ def test_saw_no_esphome_module_flag_survives_inflight_trim() -> None:
     ``test_is_no_module_named_esphome_matches_exact_quoted_form``
     above.
     """
-    from esphome_device_builder.controllers.firmware import (
-        _ERROR_PATTERNS,
+    from esphome_device_builder.controllers.firmware.constants import _ERROR_PATTERNS
+    from esphome_device_builder.controllers.firmware.helpers import (
         _is_no_module_named_esphome,
     )
