+++ tests/controllers/firmware/test_refresh.py
