+++ tests/controllers/firmware/test_rename_lock.py
