+++ tests/controllers/firmware/test_traversal_validation.py
